dependency updated

+++ src/token/api/src/state/ledger.rs
@@ -133,7 +133,7 @@ impl Ledger {
         count: usize,
         transaction_id: Option<TxId>,
     ) -> PaginatedResult {
-        let count = count as usize;
+        let count = count;
         let mut transactions = self
             .history
             .iter()
